+++ eslint.config.mjs
@@ -1,22 +1,31 @@
 import { FlatCompat } from "@eslint/eslintrc";
+import eslintPluginReact from "eslint-plugin-react";
 import { dirname } from "path";
 import { fileURLToPath } from "url";
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = dirname(__filename);
-
 const compat = new FlatCompat({
   baseDirectory: __dirname,
 });
-
 const eslintConfig = [
   ...compat.extends("next/core-web-vitals", "next/typescript"),
   {
     plugins: ["react"],
+    files: ["**/*.tsx", "**/*.jsx"],
+    languageOptions: {
+      parserOptions: {
+        ecmaFeatures: {
+          jsx: true,
+        },
+      },
+    },
+    plugins: {
+      react: eslintPluginReact,
+    },
     rules: {
       "react/no-unescaped-entities": "warn",
     },
   },
 ];
-
 export default eslintConfig;
