+++ src/components/auth/PhoneVerification.tsx
@@ -31,7 +31,7 @@ export function PhoneVerification() {
 
       router.push("/pdpa-consent");
     } catch (error) {
-      setError("Invalid verification code");
+      setError("Invalid verification code : " + { error });
     } finally {
       setIsLoading(false);
     }
@@ -47,14 +47,14 @@ export function PhoneVerification() {
         throw new Error("Failed to resend code");
       }
     } catch (error) {
-      setError("Failed to resend verification code");
+      setError("Failed to resend verification code : " + { error });
     }
   };
 
   return (
     <div className="mx-auto max-w-md space-y-4">
       <h2 className="text-center text-2xl font-bold">Verify your phone number</h2>
-      <p className="text-center text-gray-600">We've sent a verification code to your phone</p>
+      <p className="text-center text-gray-600">We have sent a verification code to your phone</p>
       <form onSubmit={onSubmit} className="space-y-4">
         <Input name="code" type="text" placeholder="Enter 6-digit code" disabled={isLoading} />
         {error && <p className="text-sm text-red-500">{error}</p>}
