+++ src/types/next-auth.d.ts
@@ -1,0 +1,22 @@
+import { DefaultJWT, DefaultSession, DefaultUser } from "next-auth";
+
+declare module "next-auth" {
+  interface Session {
+    user: DefaultSession["user"] & {
+      id: string;
+      isAdmin: boolean;
+    };
+  }
+
+  interface User extends DefaultUser {
+    id: string;
+    isAdmin: boolean;
+  }
+}
+
+declare module "next-auth/jwt" {
+  interface JWT extends DefaultJWT {
+    id: string;
+    isAdmin: boolean;
+  }
+}
